Started tracking for Oculus

+++ LichKingDungeons/TheOculus.lua
@@ -15,11 +15,22 @@ local timer
 local timerStarted = false
 
 function core.TheOculus:TrackAdditional()
-    if core.type == "UNIT_DIED" and core.destID == "27654" then
-        core:sendMessage(core:getAchievement() .. " 20 Minutes to kill Ley-Guardian Eregos")
-        if timerStarted == false then
-            timerStarted = true
+    if core.Instances[core.expansion][core.instanceType][core.instance]["boss2"].enabled == true then
+        if core.type == "UNIT_DIED" and core.destID == "27654" then
+            core:sendMessage(core:getAchievement() .. " 20 Minutes to kill Ley-Guardian Eregos")
+            if timerStarted == false then
+                timerStarted = true
+                timer = C_Timer.NewTimer(1200, function()
+                    core:sendMessage(GetAchievementLink(1868) ..  " FAILED!")
+                end) 
+            end
         end
+    
+        if core.type == "UNIT_DIED" and core.destID == "27656" then
+            if timer ~= nil then
+                timer:Cancel()
+            end
+        end   
     end
 end
